Renamed internal functions
